backend/requirements.txt
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asttokens
async-lru
attrs
babel
beautifulsoup4
bleach
certifi
cffi
charset-normalizer
colorama
comm
contourpy
curl_cffi
cycler
debugpy
decorator
defusedxml
distro
dj-rest-auth
Django
django-allauth
django-cors-headers
django-environ
djangorestframework
dotenv
et_xmlfile
executing
fastjsonschema
fonttools
fqdn
frozendict
h11
httpcore
httpx
idna
ipykernel
ipython
ipython_pygments_lexers
isoduration
jedi
Jinja2
jiter
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
kiwisolver
langchain
langchain-core
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lark
MarkupSafe
matplotlib
matplotlib-inline
mistune
multitasking
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numpy
openai
openpyxl
orjson
ormsgpack
overrides
packaging
pandas
pandocfilters
parso
peewee
pillow
platformdirs
prometheus_client
prompt_toolkit
protobuf
psutil
pure_eval
pycparser
pydantic
pydantic_core
Pygments
pyparsing
python-dateutil
python-json-logger
python-dotenv
pytz
pywin32
pywinpty
PyYAML
pyzmq
referencing
requests
requests-toolbelt
rfc3339-validator
rfc3986-validator
rfc3987-syntax
rpds-py
seaborn
Send2Trash
six
sniffio
soupsieve
sqlparse
stack-data
tenacity
terminado
tinycss2
tornado
tqdm
traitlets
types-python-dateutil
typing-inspection
typing_extensions
tzdata
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
websockets
xxhash
yfinance
zstandard